Rename CLI and runtime crates
