+++ Cargo.toml
@@ -1,14 +1,14 @@
 [workspace]
 resolver = "2"
 
-members = ["api", "cli", "examples", "macro", "parser", "runtime"]
+members = ["api", "cli", "examples", "lib", "macro", "parser"]
 
 [workspace.package]
 version = "0.16.15"
 
 [workspace.dependencies]
-kinetics = { path = "runtime", version = "0.16.15" }
 kinetics-api = { path = "api", version = "0.16.15" }
+kinetics-lib = { path = "lib", version = "0.16.15" }
 kinetics-macro = { path = "macro", version = "0.16.15" }
 kinetics-parser = { path = "parser", version = "0.16.15" }
