+++ cli/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "kinetics-cli"
+name = "kinetics"
 version.workspace = true
 edition = "2021"
 description = "Kinetics is a hosting platform for Rust applications that allows you to deploy all types of workloads by writing **only Rust code**."
@@ -40,8 +40,8 @@ keyring = { version = "3.6.3", features = [
   "vendored",
   "windows-native"
 ] }
-kinetics = { workspace = true }
 kinetics-api = { workspace = true }
+kinetics-lib = { workspace = true }
 kinetics-parser = { workspace = true }
 log = "0.4.29"
 once_cell = "1.21.4"
