+++ examples/Cargo.toml
@@ -6,7 +6,7 @@ publish = false
 
 [dependencies]
 http = "^1.0"
-kinetics = { path = "../runtime" }
+kinetics-lib = { path = "../lib" }
 matchit = "0.9.2"
 opentelemetry = { version = "0.31.0", features = ["logs", "metrics", "trace"] }
 opentelemetry-appender-tracing = { version = "0.31.1" }
