+++ examples/src/queue.rs
@@ -1,8 +1,8 @@
 use crate::basic::worker::worker;
 use http::{Request, Response, StatusCode};
-use kinetics::macros::endpoint;
-use kinetics::tools::config::Config as KineticsConfig;
-use kinetics::tools::queue::Client as QueueClient;
+use kinetics_lib::macros::endpoint;
+use kinetics_lib::tools::config::Config as KineticsConfig;
+use kinetics_lib::tools::queue::Client as QueueClient;
 use serde_json::json;
 use std::collections::HashMap;
 // As an example use a general-purpose type-erased error from tower.
