+++ lib/src/macros.rs
