+++ lib/src/tools.rs
